use uint32 in RTTInfo

+++ network/wsNetwork.go
@@ -1832,8 +1832,8 @@ func (wn *WebsocketNetwork) sendPeerConnectionsTelemetryStatus() {
 		}
 		rttInfo, err := util.GetConnRTT(peer.conn.UnderlyingConn())
 		if err == nil {
-			connDetail.RTT = rttInfo.RTT
-			connDetail.RTTVar = rttInfo.RTTVar
+			connDetail.RTT = int64(rttInfo.RTT)
+			connDetail.RTTVar = int64(rttInfo.RTTVar)
 		}
 	}
